registration

+++ database.sql
@@ -23,7 +23,7 @@ CREATE TABLE product(
 CREATE TABLE register(
     id SERIAL PRIMARY KEY,
     f_name VARCHAR(70),
-    phone SMALLINT,
+    phone INTEGER,
     email VARCHAR(100),
     password VARCHAR(100)
 );
